requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Background Tasks
celery==5.3.4
redis==5.0.1

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Google Gemini AI
google-generativeai==0.3.2

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
gunicorn==21.2.0

# Date/Time
python-dateutil==2.8.2
